style: rename hyper-sync to hypersync in filenames and imports
