+++ src/providers/evm/index.ts
@@ -1,5 +1,5 @@
 export { EvmIndexer } from './indexer';
 export { EvmProvider } from './provider';
-export { HyperSyncEvmIndexer } from './hyper-sync-indexer';
-export { HyperSyncEvmProvider } from './hyper-sync-provider';
+export { HyperSyncEvmIndexer } from './hypersync-indexer';
+export { HyperSyncEvmProvider } from './hypersync-provider';
 export * from './types';
